fix: Move autouse cleanup fixtures to integration conftest
Async autouse fixtures in the globally-registered pytest plugin
(hassette.test_utils.fixtures) were collected for sync Playwright e2e
tests, causing "Runner.run() cannot be called from a running event
loop" errors. Move all cleanup fixtures to tests/integration/conftest.py
where they're scoped to the tests that actually need them, and remove
the sync no-op override from tests/e2e/conftest.py.
